backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-genai
google-auth-oauthlib
google-api-python-client
python-dotenv
pytesseract
pillow
pypdf
pypdfium2
